tinyusb: rebase to upstream master, remove unused submodules
remove unused submodules
rebese `merge pull request #783`: Add new Espressif target esp32s3 for tinyUSB (esp-based_on_334e95f)

These submodules are not needed when TinyUSB is built as an ESP-IDF component. We remove them to avoid fetching extra submodules when ESP-IDF is cloned with --recursive flag.

Note: this is a local change, not for upstream.